Refactor InboundEventsListenerTest
